+++ Assets/scripts/Core/IOnHitEffect.cs
@@ -1,6 +1,7 @@
+using CrystalFlux.ProjectileSystem;
 using UnityEngine;
 
-namespace CrystalFlux.ProjectileSystem
+namespace CrystalFlux.Core
 {
     public interface IOnHitEffect
     {