+++ Assets/scripts/Core/IStatusEffectReceiver.cs
@@ -7,9 +7,12 @@ namespace CrystalFlux.StatusEffectSystem
     {
         void Apply(StatusEffect effect, GameObject source, Vector2 location = default);
         void ClearAllEffects();
+
         T GetActiveFirstEffectOfType<T>() where T : StatusEffect;
         void GetActiveEffectsOfType<T>(List<T> results) where T : StatusEffect;
+
         void RemoveEffectAfterDelay<T>(float delay) where T : StatusEffect;
         void RemoveEffect<T>() where T : StatusEffect;
+        void RemoveStacks<T>(int stacksToRemove) where T : StatusEffect;
     }
 }