+++ Assets/scripts/Entity/Player/PlayerAttackCooldownUI.cs
@@ -98,7 +98,7 @@ namespace CrystalFlux.EntitySystem
             if (cad.pd != null)
             {
                 var previewSnapshot = ProjectileSnapshot.CaptureSnapshot(cad.pd, (cesm as Component).gameObject);
-                var previewPacket = DamagePacket.BuildDamagePacket(cad.pd, previewSnapshot, false, (cesm as Component).gameObject, false, 1f);
+                var previewPacket = DamagePacketBuilder.BuildDamagePacket(cad.pd, previewSnapshot, false, (cesm as Component).gameObject, false, 1f);
 
                 foreach (var instance in previewPacket.instances)
                 {