+++ Assets/scripts/Projectile/DamagePacketBuilder.cs
@@ -1,30 +1,11 @@
-using System.Collections.Generic;
+
 using CrystalFlux.Core;
 using UnityEngine;
 
 namespace CrystalFlux.ProjectileSystem
 {
-    public class DamagePacket
+    public static class DamagePacketBuilder
     {
-        public List<DamageInstance> instances = new();
-        public GameObject source;
-        public bool bypassIFrames = false;
-        public float sizeOverride = 1f;
-
-        public void AddInstance(DamageType type, float amount, bool isCrit, GameObject owner)
-            => instances.Add(new DamageInstance(type, amount, isCrit, default, owner));
-
-        public void AddInstance(DamageType type, float amount, bool isCrit, Color indicatorColor, GameObject owner)
-            => instances.Add(new DamageInstance(type, amount, isCrit, indicatorColor, owner));
-
-        public float GetTotalDamage()
-        {
-            float total = 0f;
-            foreach (var i in instances)
-                total += i.amount;
-            return total;
-        }
-
         public static DamagePacket BuildDamagePacket(ProjectileData pd, ProjectileDamageSnapshot snapshot, bool rollCrits, GameObject owner, bool bypassIFrames, float sizeOverride)
         {
             DamagePacket dp = new() { source = owner, bypassIFrames = bypassIFrames, sizeOverride = sizeOverride };